pyproject.toml
[project]
name = "ptt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "aiortc>=1.10.1",
    "pyserial>=3.5",
]

[tool.uv.workspace]
members = [
    "ptt-box",
]
